backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0